+++ apps/wallets/quickstart-devkit/app/providers.tsx
@@ -1,10 +1,12 @@
 "use client";
 
+import { Suspense } from "react";
 import { CrossmintAuthProvider, CrossmintProvider, CrossmintWalletProvider } from "@crossmint/client-sdk-react-ui";
 import { EthereumWalletConnectors } from "@dynamic-labs/ethereum";
 import { DynamicContextProvider } from "@dynamic-labs/sdk-react-core";
 import { SolanaWalletConnectors } from "@dynamic-labs/solana";
 import { PrivyProvider } from "@privy-io/react-auth";
+import { useSearchParams } from "next/navigation";
 
 const crossmintApiKey = process.env.NEXT_PUBLIC_CROSSMINT_API_KEY ?? "";
 if (!crossmintApiKey) {
@@ -12,30 +14,43 @@ if (!crossmintApiKey) {
 }
 
 export function Providers({ children }: { children: React.ReactNode }) {
-    /* @TODO update to your desired provider here */
-    return <EVMCrossmintAuthProvider>{children}</EVMCrossmintAuthProvider>;
+    /*
+    @TODO update to your desired provider inside QueryParamsProvider.
+    (Ignore this, it's used for e2e testing - Do not remove this)
+    */
+    return (
+        <Suspense>
+            <QueryParamsProvider>{children}</QueryParamsProvider>
+        </Suspense>
+    );
 }
 
 /* ============================================================ */
 /*                    ALL EVM WALLET PROVIDERS                  */
 /* ============================================================ */
-function EVMCrossmintAuthProvider({ children }: { children: React.ReactNode }) {
+function EVMCrossmintAuthProvider({
+    children,
+    apiKey,
+    createOnLogin,
+}: { children: React.ReactNode; apiKey?: string; createOnLogin?: any }) {
     if (!process.env.NEXT_PUBLIC_EVM_CHAIN) {
         console.error("NEXT_PUBLIC_EVM_CHAIN is not set");
         return;
     }
+
     return (
-        <CrossmintProvider apiKey={process.env.NEXT_PUBLIC_CROSSMINT_API_KEY || ""}>
+        <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
             <CrossmintAuthProvider
                 authModalTitle="EVM Wallets Quickstart"
                 loginMethods={["google", "twitter", "email", "web3:evm-only"]}
             >
                 <CrossmintWalletProvider
                     showPasskeyHelpers={false}
-                    createOnLogin={{
-                        chain: process.env.NEXT_PUBLIC_EVM_CHAIN as any,
-                        signer: { type: "external-wallet" },
-                    }}
+                    createOnLogin={
+                        createOnLogin != null
+                            ? createOnLogin
+                            : { chain: process.env.NEXT_PUBLIC_EVM_CHAIN as any, signer: { type: "email" } }
+                    }
                 >
                     {children}
                 </CrossmintWalletProvider>
@@ -44,7 +59,7 @@ function EVMCrossmintAuthProvider({ children }: { children: React.ReactNode }) {
     );
 }
 
-function EVMPrivyProvider({ children }: { children: React.ReactNode }) {
+function EVMPrivyProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     if (!process.env.NEXT_PUBLIC_PRIVY_APP_ID) {
         throw new Error("NEXT_PUBLIC_PRIVY_APP_ID is not set");
     }
@@ -60,14 +75,14 @@ function EVMPrivyProvider({ children }: { children: React.ReactNode }) {
                 },
             }}
         >
-            <CrossmintProvider apiKey={crossmintApiKey}>
+            <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
                 <CrossmintWalletProvider showPasskeyHelpers={false}>{children}</CrossmintWalletProvider>
             </CrossmintProvider>
         </PrivyProvider>
     );
 }
 
-function EVMDynamicLabsProvider({ children }: { children: React.ReactNode }) {
+function EVMDynamicLabsProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     if (!process.env.NEXT_PUBLIC_DYNAMIC_ENV_ID) {
         throw new Error("NEXT_PUBLIC_DYNAMIC_ENV_ID is not set");
     }
@@ -78,20 +93,20 @@ function EVMDynamicLabsProvider({ children }: { children: React.ReactNode }) {
                 walletConnectors: [EthereumWalletConnectors],
             }}
         >
-            <CrossmintProvider apiKey={crossmintApiKey}>
+            <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
                 <CrossmintWalletProvider>{children}</CrossmintWalletProvider>
             </CrossmintProvider>
         </DynamicContextProvider>
     );
 }
 
-function EVMFirebaseProvider({ children }: { children: React.ReactNode }) {
+function EVMFirebaseProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     if (!process.env.NEXT_PUBLIC_FIREBASE_API_KEY) {
         console.error("Make sure to set all firebase .env vars for Firebase BYOA");
         return;
     }
     return (
-        <CrossmintProvider apiKey={crossmintApiKey}>
+        <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
             <CrossmintWalletProvider>{children}</CrossmintWalletProvider>
         </CrossmintProvider>
     );
@@ -100,16 +115,22 @@ function EVMFirebaseProvider({ children }: { children: React.ReactNode }) {
 /* ============================================================ */
 /*                    ALL SOLANA WALLET PROVIDERS               */
 /* ============================================================ */
-function SolanaCrossmintAuthProvider({ children }: { children: React.ReactNode }) {
+function SolanaCrossmintAuthProvider({
+    children,
+    apiKey,
+    createOnLogin,
+}: { children: React.ReactNode; apiKey?: string; createOnLogin?: any }) {
     return (
-        <CrossmintProvider apiKey={process.env.NEXT_PUBLIC_CROSSMINT_API_KEY || ""}>
+        <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
             <CrossmintAuthProvider
                 authModalTitle="Solana Wallets Quickstart"
                 loginMethods={["google", "twitter", "web3:solana-only", "email"]}
             >
                 <CrossmintWalletProvider
                     showPasskeyHelpers={false}
-                    createOnLogin={{ chain: "solana", signer: { type: "email" } }}
+                    createOnLogin={
+                        createOnLogin != null ? createOnLogin : { chain: "solana", signer: { type: "email" } }
+                    }
                 >
                     {children}
                 </CrossmintWalletProvider>
@@ -118,7 +139,7 @@ function SolanaCrossmintAuthProvider({ children }: { children: React.ReactNode }
     );
 }
 
-function SolanaPrivyProvider({ children }: { children: React.ReactNode }) {
+function SolanaPrivyProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     if (!process.env.NEXT_PUBLIC_PRIVY_APP_ID) {
         throw new Error("NEXT_PUBLIC_PRIVY_APP_ID is not set");
     }
@@ -134,7 +155,7 @@ function SolanaPrivyProvider({ children }: { children: React.ReactNode }) {
                 },
             }}
         >
-            <CrossmintProvider apiKey={crossmintApiKey}>
+            <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
                 <CrossmintWalletProvider showPasskeyHelpers={false}>{children}</CrossmintWalletProvider>
             </CrossmintProvider>
         </PrivyProvider>
@@ -143,8 +164,10 @@ function SolanaPrivyProvider({ children }: { children: React.ReactNode }) {
 
 function SolanaDynamicLabsProvider({
     children,
+    apiKey,
 }: {
     children: React.ReactNode;
+    apiKey?: string;
 }) {
     if (!process.env.NEXT_PUBLIC_DYNAMIC_ENV_ID) {
         throw new Error("NEXT_PUBLIC_DYNAMIC_ENV_ID is not set");
@@ -156,24 +179,24 @@ function SolanaDynamicLabsProvider({
                 walletConnectors: [SolanaWalletConnectors],
             }}
         >
-            <CrossmintProvider apiKey={crossmintApiKey}>
+            <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
                 <CrossmintWalletProvider>{children}</CrossmintWalletProvider>
             </CrossmintProvider>
         </DynamicContextProvider>
     );
 }
 
-function SolanaFirebaseProvider({ children }: { children: React.ReactNode }) {
+function SolanaFirebaseProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     return (
-        <CrossmintProvider apiKey={crossmintApiKey}>
+        <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
             <CrossmintWalletProvider>{children}</CrossmintWalletProvider>
         </CrossmintProvider>
     );
 }
 
-function StellarCrossmintAuthProvider({ children }: { children: React.ReactNode }) {
+function StellarCrossmintAuthProvider({ children, apiKey }: { children: React.ReactNode; apiKey?: string }) {
     return (
-        <CrossmintProvider apiKey={process.env.NEXT_PUBLIC_CROSSMINT_API_KEY || ""}>
+        <CrossmintProvider apiKey={apiKey ?? crossmintApiKey}>
             <CrossmintAuthProvider
                 authModalTitle="Stellar Wallets Quickstart"
                 loginMethods={["google", "twitter", "email", "web3"]}
@@ -190,4 +213,63 @@ function StellarCrossmintAuthProvider({ children }: { children: React.ReactNode 
             </CrossmintAuthProvider>
         </CrossmintProvider>
     );
+}
+
+// This provider is used for testing and development purposes
+// It allows you to specify the provider, chain, signer, and chainId via URL params
+function QueryParamsProvider({ children }: { children: React.ReactNode }) {
+    const searchParams = useSearchParams();
+
+    const providerType = searchParams.get("provider") || "crossmint"; // default to crossmint
+    const chainType = searchParams.get("chain");
+    const signerType = searchParams.get("signer");
+    const chainId = searchParams.get("chainId") || process.env.NEXT_PUBLIC_EVM_CHAIN;
+    const phoneNumber = searchParams.get("phoneNumber");
+    const crossmintApiKey = searchParams.get("crossmintApiKey") || undefined;
+
+    if (chainType === "evm") {
+        switch (providerType) {
+            case "privy":
+                return <EVMPrivyProvider>{children}</EVMPrivyProvider>;
+            case "dynamic":
+                return <EVMDynamicLabsProvider>{children}</EVMDynamicLabsProvider>;
+            case "firebase":
+                return <EVMFirebaseProvider>{children}</EVMFirebaseProvider>;
+            case "crossmint":
+            default:
+                const createOnLogin: any = { chain: chainId, signer: { type: signerType } };
+                if (signerType === "phone" && phoneNumber != null) {
+                    createOnLogin.signer = { type: signerType, phone: decodeURIComponent(phoneNumber) };
+                }
+                return (
+                    <EVMCrossmintAuthProvider apiKey={crossmintApiKey} createOnLogin={createOnLogin}>
+                        {children}
+                    </EVMCrossmintAuthProvider>
+                );
+        }
+    } else if (chainType === "solana") {
+        switch (providerType) {
+            case "privy":
+                return <SolanaPrivyProvider>{children}</SolanaPrivyProvider>;
+            case "dynamic":
+                return <SolanaDynamicLabsProvider>{children}</SolanaDynamicLabsProvider>;
+            case "firebase":
+                return <SolanaFirebaseProvider>{children}</SolanaFirebaseProvider>;
+            case "crossmint":
+            default:
+                const createOnLogin: any = { chain: "solana", signer: { type: signerType } };
+                if (signerType === "phone" && phoneNumber != null) {
+                    createOnLogin.signer = { type: signerType, phone: decodeURIComponent(phoneNumber) };
+                }
+                return (
+                    <SolanaCrossmintAuthProvider apiKey={crossmintApiKey} createOnLogin={createOnLogin}>
+                        {children}
+                    </SolanaCrossmintAuthProvider>
+                );
+        }
+    } else if (chainType === "stellar") {
+        return <StellarCrossmintAuthProvider>{children}</StellarCrossmintAuthProvider>;
+    }
+
+    return <EVMCrossmintAuthProvider>{children}</EVMCrossmintAuthProvider>;
 }