+++ packages/client/base/src/utils/i18n/dictionaries/esES.ts
@@ -1,11 +1,4 @@
 const esES = {
-    crossmintPayButtonService: {
-        CONNECTING: "Conectando...",
-        BUY: "Comprar con Crossmint",
-        BUY_WITH_ETH: "Comprar con ETH",
-        BUY_WITH_SOL: "Comprar con SOL",
-        BUY_WITH_CREDIT_CARD: "Comprar con tarjeta de crédito",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "Pagar con {0}",