+++ packages/client/base/src/utils/i18n/dictionaries/viVN.ts
@@ -1,10 +1,4 @@
 const enUS = {
-    crossmintPayButtonService: {
-        CONNECTING: "Đang kết nối",
-        BUY_WITH_ETH: "Mua bằng ETH",
-        BUY_WITH_SOL: "Mua bằng SOL",
-        BUY_WITH_CREDIT_CARD: "Mua bằng thẻ tín dụng",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "Thanh toán bằng {0}",