+++ packages/client/ui/react-native/src/index.ts
@@ -12,6 +12,7 @@ export {
     type Chain,
     type DelegatedSigner,
     type Transaction,
+    type Signature,
     EVMWallet,
     SolanaWallet,
     Wallet,