+++ packages/client/ui/react-ui/src/components/hosted/index.ts
@@ -1,3 +1,1 @@
-export * from "./CrossmintPayButton_DEPRECATED";
-
 export * from "./v3";