+++ packages/client/ui/react-ui/src/providers/CrossmintWalletProvider.tsx
@@ -155,7 +155,6 @@ export function CrossmintWalletProvider({
         try {
             await verifyOtpRef.current(otp);
             setEmailSignerDialogOpen(false);
-            setEmailSignerDialogStep("initial");
         } catch (error) {
             console.error("Failed to verify OTP", error);
             rejectRef.current(new Error("Failed to verify OTP"));
@@ -177,7 +176,6 @@ export function CrossmintWalletProvider({
         try {
             await verifyPhoneOtpRef.current(otp);
             setPhoneSignerDialogOpen(false);
-            setPhoneSignerDialogStep("initial");
         } catch (error) {
             console.error("Failed to verify phone OTP", error);
             rejectRef.current(new Error("Failed to verify phone OTP"));