+++ packages/client/ui/vanilla-ui/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./components";
-
-export { CrossmintEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
-export type { CrossmintEvent, CrossmintEventMap } from "@crossmint/client-sdk-base";