+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # crossmint-auth-node
 
+## 1.1.51
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.40
+
+## 1.1.50
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.39
+
 ## 1.1.49
 
 ### Patch Changes