+++ apps/wallets/quickstart-devkit/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @crossmint/wallets-quickstart-devkit
 
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies [40917df]
+  - @crossmint/client-sdk-react-ui@2.4.0
+
+## 0.1.46
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.3.4
+
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies [57e4202]
+- Updated dependencies [4728def]
+  - @crossmint/client-sdk-react-ui@2.3.3
+
+## 0.1.44
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.3.2
+
 ## 0.1.43
 
 ### Patch Changes