+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.158
+
+### Patch Changes
+
+- Updated dependencies [40917df]
+  - @crossmint/client-sdk-react-ui@2.4.0
+
+## 0.1.157
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.3.4
+
+## 0.1.156
+
+### Patch Changes
+
+- Updated dependencies [57e4202]
+- Updated dependencies [4728def]
+  - @crossmint/client-sdk-react-ui@2.3.3
+
+## 0.1.155
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.3.2
+
 ## 0.1.154
 
 ### Patch Changes